cloud_service/requirements.txt

flask
ultralytics
pillow
gunicorn
firebase-admin
google-cloud-firestore